ref: set up folder structure and package jsons
