+++ client/src/App.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function App() {
+  return (
+    <div>Hello World</div>
+  );
+}
+
+export default App;
